# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KV Silver Zone API",
    "version": "1.0.0",
    "description": "API documentation for KV Silver Zone. Auth is via an httpOnly cookie set on /auth/login (a Bearer token is also returned for API clients). Protected endpoints accept either."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error / bad input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict (e.g. duplicate or insufficient stock)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "price must be a positive number"
          }
        }
      },
      "ProductImage": {
        "type": "object",
        "properties": {
          "variantName": {
            "type": "string",
            "example": "Default view"
          },
          "imageBase64": {
            "type": "string",
            "description": "Base64 data URL or image URL"
          },
          "sortOrder": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "ProductPricing": {
        "type": "object",
        "description": "Live pricing breakdown computed server-side from the silver rate.",
        "properties": {
          "basis": {
            "type": "string",
            "enum": [
              "live",
              "static"
            ],
            "example": "live"
          },
          "metalValue": {
            "type": "number",
            "example": 925
          },
          "makingCharge": {
            "type": "number",
            "example": 92.5
          },
          "ratePerGram": {
            "type": "number",
            "nullable": true,
            "example": 100
          },
          "purityFraction": {
            "type": "number",
            "example": 0.925
          },
          "currency": {
            "type": "string",
            "example": "INR"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "665f1c2a9b1e4a0012ab34cd"
          },
          "productGroupCode": {
            "type": "string",
            "example": "RING001"
          },
          "name": {
            "type": "string",
            "example": "Silver Lotus Ring"
          },
          "description": {
            "type": "string"
          },
          "material": {
            "type": "string",
            "example": "Silver"
          },
          "weight": {
            "type": "number",
            "example": 10
          },
          "weightInGrams": {
            "type": "number",
            "example": 10
          },
          "price": {
            "type": "number",
            "description": "Authoritative live price",
            "example": 1017.5
          },
          "listedPrice": {
            "type": "number",
            "description": "Static fallback price",
            "example": 5000
          },
          "originalPrice": {
            "type": "number"
          },
          "purity": {
            "type": "string",
            "example": "925"
          },
          "makingChargePercent": {
            "type": "number",
            "example": 10
          },
          "makingChargePerGram": {
            "type": "number"
          },
          "isSale": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean",
            "description": "Catalog visibility (admin-controlled)"
          },
          "quantity": {
            "type": "integer",
            "description": "Initial/seed quantity"
          },
          "stockAvailable": {
            "type": "integer",
            "description": "Live stock from Inventory",
            "example": 7
          },
          "inStock": {
            "type": "boolean",
            "example": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImage"
            }
          },
          "pricing": {
            "$ref": "#/components/schemas/ProductPricing"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": [
          "name",
          "category",
          "weight",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Silver Lotus Ring"
          },
          "category": {
            "type": "string",
            "description": "Stored as material",
            "example": "Silver"
          },
          "weight": {
            "type": "number",
            "example": 10
          },
          "price": {
            "type": "number",
            "example": 5000
          },
          "quantity": {
            "type": "integer",
            "example": 10
          },
          "description": {
            "type": "string"
          },
          "purity": {
            "type": "string",
            "example": "925"
          },
          "originalPrice": {
            "type": "number"
          },
          "makingChargePercent": {
            "type": "number",
            "example": 10
          },
          "makingChargePerGram": {
            "type": "number"
          },
          "image": {
            "type": "string",
            "description": "Base64 data URL or image URL"
          },
          "productGroupCode": {
            "type": "string",
            "description": "Auto-generated from name if omitted"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "isSale": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReviewInput": {
        "type": "object",
        "required": [
          "rating"
        ],
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/products": {
      "post": {
        "summary": "Add new product",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "summary": "Get all products with filtering",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category"
          },
          {
            "in": "query",
            "name": "metal",
            "schema": {
              "type": "string"
            },
            "description": "Filter by metal/purity"
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "number"
            },
            "description": "Minimum price"
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "number"
            },
            "description": "Maximum price"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Full-text search term"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "price_asc",
                "price_desc",
                "newest"
              ]
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "onSale",
            "schema": {
              "type": "boolean"
            },
            "description": "Only products on sale"
          },
          {
            "in": "query",
            "name": "featured",
            "schema": {
              "type": "boolean"
            },
            "description": "Only featured products"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/products/{id}": {
      "put": {
        "summary": "Update product",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete product",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "get": {
        "summary": "Get product by ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/products/{productId}/reviews": {
      "get": {
        "summary": "Get all reviews for a product",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews for the product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "summary": "Submit a review for a product",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/products/{productId}/reviews/{reviewId}": {
      "delete": {
        "summary": "Delete a review (owner or admin)",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
...
}